move courses collection to /imports
