+++ collections/events.js
@@ -1,5 +1,7 @@
 import '/imports/Filtering.js';
 import '/imports/Predicates.js';
+import Courses from '/imports/api/courses/courses.js';
+import UpdateMethods from '/imports/utils/update-methods.js';
 
 import '/imports/notification/Notification.js';
 import '/imports/LocalTime.js';
