+++ imports/notification/Notification.Join.js
@@ -1,6 +1,7 @@
 export default notificationJoin = {};
 import '/imports/collections/Log.js';
 import '/imports/StringTools.js';
+import Courses from '/imports/api/courses/courses.js';
 
 /** Record the intent to send join notifications
   *
