+++ imports/startup/server/fixtures.js
@@ -1,4 +1,5 @@
 import '/imports/api/fixtures/methods.js';
+import Courses from '/imports/api/courses/courses.js';
 
 if (Meteor.settings.testdata) {
 	const logResult = function(error, result) {
