+++ imports/startup/server/index.js
@@ -1,4 +1,5 @@
 import './api.json.js';
 import './fixtures.js';
 import './notifications.js';
+import './register-api.js';
 import './login.js';
