+++ imports/startup/server/register-api.js
@@ -1,0 +1,2 @@
+import '/imports/api/courses/methods.js';
+import '/imports/api/courses/server/publications.js';
