+++ imports/ui/components/profiles/course-list/profile-course-list.js
@@ -1,4 +1,5 @@
 import { Template } from 'meteor/templating';
+import Courses from '/imports/api/courses/courses.js';
 
 import ScssVars from '/imports/ui/lib/scss-vars.js';
