+++ imports/ui/pages/frames/courselist/courselist-frame.js
@@ -4,6 +4,7 @@ import { Session } from 'meteor/session';
 import { Template } from 'meteor/templating';
 import { $ } from 'meteor/jquery';
 
+import Courses from '/imports/api/courses/courses.js';
 import Metatags from '/imports/Metatags.js';
 import CssFromQuery from '/imports/ui/lib/css-from-query.js';
