+++ server/extracts.msgfmt~
@@ -1,1 +1,1 @@
-[{"loginAction.enroll":{"key":"loginAction.enroll","text":"Login and enroll","file":"imports/ui/components/courses/roles/course-roles.js","line":57,"mtime":1511526909043,"template":"unknown","ctime":1511526909043},"loginAction.saveVenue":{"key":"loginAction.saveVenue","text":"Login and save venue","file":"imports/ui/components/venues/edit/venue-edit.js","line":194,"mtime":1511526897290,"template":"unknown","ctime":1511526897290},"venue.create.plsGiveVenueName":{"key":"venue.create.plsGiveVenueName","text":"Please give your venue a name","file":"imports/ui/components/venues/edit/venue-edit.js","line":158,"mtime":1511526897290,"template":"unknown"},"venue.create.plsProvideDescription":{"key":"venue.create.plsProvideDescription","text":"Please provide a description for your venue","file":"imports/ui/components/venues/edit/venue-edit.js","line":166,"mtime":1511526897290,"template":"unknown"},"venue.create.plsSelectPointOnMap":{"key":"venue.create.plsSelectPointOnMap","text":"Please select a point on the map","file":"imports/ui/components/venues/edit/venue-edit.js","line":188,"mtime":1511526897290,"template":"unknown"},"venue.create.plsSelectRegion":{"key":"venue.create.plsSelectRegion","text":"Please select a region","file":"imports/ui/components/venues/edit/venue-edit.js","line":179,"mtime":1511526897290,"template":"unknown"},"venue.edit.description.placeholder":{"key":"venue.edit.description.placeholder","text":"Some words about this venue","file":"imports/ui/components/venues/edit/venue-edit.js","line":79,"mtime":1511526897290,"template":"unknown"},"venue.saving.success":{"key":"venue.saving.success","text":"Saved changes to venue \"{NAME}\".","file":"imports/ui/components/venues/edit/venue-edit.js","line":200,"mtime":1511526897290,"template":"unknown"}},{"extractedAt":1511526909701,"updatedAt":1511526909043}]
+[{"courses.creator.defaultMessage":{"key":"courses.creator.defaultMessage","text":"(has proposed this course)","file":"imports/api/courses/methods.js","line":337,"mtime":1511699009756,"template":"unknown","ctime":1511699009756}},{"extractedAt":1511699022674,"updatedAt":1511699009756}]
