+++ server/updates/2016.03.29 ensureInternalField.js
@@ -1,3 +1,5 @@
+import Courses from '/imports/api/courses/courses.js';
+
 // Ensure no null groups in events
 UpdatesAvailable.ensureInternalField = function() {
 	return Events.update({ internal: null }, { $set: { internal: false } }, { multi: true })
