+++ server/updates/2016.04.26 ensureGroupOrganizersField.js
@@ -1,3 +1,5 @@
+import Courses from '/imports/api/courses/courses.js';
+
 UpdatesAvailable.ensureGroupOrganizersField = function() {
 	return Courses.update({ groupOrganizers: null }, { $set: { groupOrganizers: [] } }, { multi: true });
 };
